Renaming around hook creation code
